backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
anthropic>=0.20.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
# Supabase — PostgreSQL driver + Python client
psycopg2-binary==2.9.9
supabase>=2.3.0
